+++ CHANGELOG.md
@@ -8,8 +8,8 @@
 
 ### Improvements
 - Better error handling using ErrorBoundary, allow user self recover from error state
-
 - Code rebranding, replace `Odin` in code
+- Add sentry integration
 
 # v0.9.2 (2019.04.09)